+++ ui/src/pages/namespace/Settings/Variables/MimeTypeSelect/index.tsx
@@ -1,3 +1,4 @@
+import { EditorMimeTypeSchema, MimeTypeType, mimeTypes } from "../utils";
 import {
   Select,
   SelectContent,
@@ -7,47 +8,6 @@ import {
 } from "~/design/Select";
 
 import { useTranslation } from "react-i18next";
-import { z } from "zod";
-
-const mimeTypes = [
-  { label: "JSON", value: "application/json" },
-  { label: "YAML", value: "application/yaml" },
-  { label: "shell", value: "application/x-sh" },
-  { label: "plaintext", value: "text/plain" },
-  { label: "HTML", value: "text/html" },
-  { label: "CSS", value: "text/css" },
-];
-
-export const mimeTypeToLanguageDict = {
-  "application/json": "json",
-  "application/yaml": "yaml",
-  "application/x-sh": "shell",
-  "text/plain": "plaintext",
-  "text/html": "html",
-  "text/css": "css",
-} as const;
-
-export const getLanguageFromMimeType = (mimeType: string) => {
-  const parsed = EditorMimeTypeSchema.safeParse(mimeType);
-  if (parsed.success) {
-    return mimeTypeToLanguageDict[parsed.data];
-  }
-  return undefined;
-};
-
-export const MimeTypeSchema = z.string();
-export type MimeTypeType = z.infer<typeof MimeTypeSchema>;
-
-export const EditorMimeTypeSchema = z.enum([
-  "application/json",
-  "application/yaml",
-  "application/x-sh",
-  "text/plain",
-  "text/html",
-  "text/css",
-]);
-
-export type TextMimeTypeType = z.infer<typeof EditorMimeTypeSchema>;
 
 const MimeTypeSelect = ({
   id,
